+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/DriveConstants.java
@@ -92,6 +92,9 @@ public class DriveConstants {
     public static double MAX_ANG_VEL = Math.toRadians(63.0254);
     public static double MAX_ANG_ACCEL = Math.toRadians(63.0254);
 
+    public double getMaxVel(){
+    return MAX_VEL;
+    }
 
     public static double encoderTicksToInches(double ticks) {
         return WHEEL_RADIUS * 2 * Math.PI * GEAR_RATIO * ticks / TICKS_PER_REV;
