+++ settings.gradle
@@ -1,3 +1,3 @@
 include ':FtcRobotController'
 include ':TeamCode'
-include ':MeepMeepTesting'
+include ':MeepMeepPathVisualizer'
